client->user:over ,wait distribute call

+++ internal/model/distribution.go
@@ -7,8 +7,8 @@ import (
 type Distribution struct {
 	ggorm.Model `gorm:"embedded"` // embed id and time
 
-	UserID      int `gorm:"column:user_id;type:int(11)" json:"userId"`
-	GroupCallID int `gorm:"column:group_call_id;type:int(11)" json:"groupCallId"`
+	UserID      int    `gorm:"column:user_id;type:int(11)" json:"userId"`
+	GroupCallID uint64 `gorm:"column:group_call_id;type:int(11)" json:"groupCallId"`
 }
 
 // TableName table name
